requirements.txt
alabaster==0.7.12
amqp==1.4.9
anyjson==0.3.3
argon2-cffi==18.1.0
asgiref==3.2.7
attrs==19.3.0
awesome-slugify==1.6.5
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.6.0
billiard==3.3.0.23
celery==3.1.25
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.16
django-allauth==0.35.0
django-cors-headers==3.5.0
django-coverage-plugin==1.5.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-extensions==2.0.6
django-filter==2.2.0
django-import-export==2.1.0
django-model-utils==3.1.1
django-nose==1.4.7
django-redis==4.9.0
django-test-plus==1.0.22
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
docutils==0.16
drf-yasg==1.4.4
environ==1.0
et-xmlfile==1.0.1
factory-boy==2.10.0
Faker==4.1.1
flake8==3.5.0
flex==6.12.0
future==0.16.0
idna==2.5
imagesize==1.2.0
importlib-metadata==1.6.0
inflection==0.5.0
ipdb==0.11
ipython==7.9.0
ipython-genutils==0.2.0
itypes==1.2.0
jdcal==1.4.1
jedi==0.17.1
Jinja2==2.11.2
jsondiff==1.1.1
jsonfield==3.0.0
jsonpointer==2.0
jsonschema==3.2.0
kombu==3.0.37
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
nose==1.3.7
oauthlib==3.1.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==2.6.4
packaging==20.3
parso==0.7.0
pathlib2==2.3.5
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.0.0
pluggy==0.6.0
prompt-toolkit==2.0.10
ptyprocess==0.6.0
py==1.8.1
pycodestyle==2.3.1
pycparser==2.20
pyflakes==1.6.0
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==3.5.0
pytest-django==3.1.2
pytest-sugar==0.9.1
python-dateutil==2.6.1
python3-openid==3.2.0
pytz==2018.3
PyYAML==5.3.1
redis==3.5.3
regex==2020.6.8
requests==2.18.1
requests-oauthlib==1.3.0
rfc3987==1.3.8
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.11.0
snowballstemmer==2.0.0
Sphinx==1.7.2
sphinxcontrib-websupport==1.2.3
sqlparse==0.3.1
stevedore==1.32.0
strict-rfc3339==0.7
swagger-spec-validator==2.1.0
tablib==1.1.0
termcolor==1.1.0
text-unidecode==1.3
traitlets==4.3.2
Unidecode==0.4.21
unittest-tap-reporting==0.2
unittest-xml-reporting==2.1.0
uritemplate==3.0.1
urllib3==1.21.1
validate-email==1.3
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
wcwidth==0.1.9
Werkzeug==0.14.1
xlrd==1.2.0
xlwt==1.3.0
zipp==1.2.0
